refactor: remove kitten references
